Rename ColorService #337
