--- visualization/app/codeCharta/core/color/color.module.ts
@@ -1,9 +1,0 @@
-import { ColorService } from "./color.service";
-import angular from "angular";
-
-angular.module("app.codeCharta.core.color", [])
-    .service(
-        "colorService", ColorService
-    );
-
-
